Eliminated some warnings/errors

+++ app/src/main/java/com/example/workflow/CalendarAdapter.java
@@ -32,7 +32,7 @@ class CalendarAdapter extends RecyclerView.Adapter<CalendarViewHolder>
         if(days.size() > 15) //month view
             layoutParams.height = (int) (parent.getHeight() * 0.166666666);
         else // week view
-            layoutParams.height = (int) parent.getHeight();
+            layoutParams.height = parent.getHeight();
 
         return new CalendarViewHolder(view, onItemListener, days);
     }
